requirements.txt
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0
opencv-python>=4.5.0
pillow>=8.0.0
matplotlib>=3.3.0
seaborn>=0.11.0
scikit-learn>=0.24.0
tqdm>=4.60.0
numpy>=1.21.0
pandas>=1.3.0

# Web Framework
django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# Pose Estimation
mediapipe>=0.10.0

# Development and Testing
python-decouple>=3.8 